organise electron code
