+++ electron/src/get-keytar.ts
